+++ remote/test/puppeteer/packages/puppeteer-core/src/cdp/Connection.ts
@@ -18,17 +18,13 @@ import type {ConnectionTransport} from '../common/ConnectionTransport.js';
 import {debug} from '../common/Debug.js';
 import {TargetCloseError} from '../common/Errors.js';
 import {EventEmitter} from '../common/EventEmitter.js';
+import {assert} from '../util/assert.js';
 import {createProtocolErrorMessage} from '../util/ErrorLike.js';
 
 import {CdpCDPSession} from './CDPSession.js';
 
 const debugProtocolSend = debug('puppeteer:protocol:SEND ►');
 const debugProtocolReceive = debug('puppeteer:protocol:RECV ◀');
-
-/**
- * @public
- */
-export type {ConnectionTransport, ProtocolMapping};
 
 /**
  * @public
@@ -61,6 +57,13 @@ export class Connection extends EventEmitter<CDPSessionEvents> {
 
   static fromSession(session: CDPSession): Connection | undefined {
     return session.connection();
+  }
+
+  /**
+   * @internal
+   */
+  get delay(): number {
+    return this.#delay;
   }
 
   get timeout(): number {
@@ -117,6 +120,8 @@ export class Connection extends EventEmitter<CDPSessionEvents> {
     sessionId?: string,
     options?: CommandOptions
   ): Promise<ProtocolMapping.Commands[T]['returnType']> {
+    assert(!this.#closed, 'Protocol error: Connection closed.');
+
     return callbacks.create(method, options?.timeout ?? this.#timeout, id => {
       const stringifiedMessage = JSON.stringify({
         method,