+++ remote/test/puppeteer/packages/puppeteer-core/src/revisions.ts
@@ -8,7 +8,7 @@
  * @internal
  */
 export const PUPPETEER_REVISIONS = Object.freeze({
-  chrome: '123.0.6312.122',
-  'chrome-headless-shell': '123.0.6312.122',
+  chrome: '125.0.6422.60',
+  'chrome-headless-shell': '125.0.6422.60',
   firefox: 'latest',
 });