+++ remote/test/puppeteer/test/src/requestinterception-experimental.spec.ts
@@ -106,7 +106,6 @@ describe('cooperative request interception', function () {
           expect(request.method()).toBe('GET');
           expect(request.postData()).toBe(undefined);
           expect(request.isNavigationRequest()).toBe(true);
-          expect(request.resourceType()).toBe('document');
           expect(request.frame()!.url()).toBe('about:blank');
           expect(request.frame() === page.mainFrame()).toBe(true);
         } catch (error) {
@@ -141,7 +140,7 @@ describe('cooperative request interception', function () {
       `);
       await Promise.all([
         page.$eval('form', form => {
-          return (form as HTMLFormElement).submit();
+          return form.submit();
         }),
         page.waitForNavigation(),
       ]);
@@ -408,7 +407,6 @@ describe('cooperative request interception', function () {
       expect(response!.status()).toBe(200);
       expect(response!.url()).toContain('empty.html');
       expect(requests).toHaveLength(5);
-      expect(requests[2]!.resourceType()).toBe('document');
       // Check redirect chain
       const redirectChain = response!.request().redirectChain();
       expect(redirectChain).toHaveLength(4);
@@ -442,8 +440,6 @@ describe('cooperative request interception', function () {
       expect(response!.status()).toBe(200);
       expect(response!.url()).toContain('one-style.html');
       expect(requests).toHaveLength(5);
-      expect(requests[0]!.resourceType()).toBe('document');
-      expect(requests[1]!.resourceType()).toBe('stylesheet');
       // Check redirect chain
       const redirectChain = requests[1]!.redirectChain();
       expect(redirectChain).toHaveLength(3);
@@ -631,7 +627,7 @@ describe('cooperative request interception', function () {
       void (page.$eval(
         'iframe',
         (frame, url) => {
-          return ((frame as HTMLIFrameElement).src = url as string);
+          return (frame.src = url as string);
         },
         server.EMPTY_PAGE
       ),
@@ -1004,6 +1000,38 @@ describe('cooperative request interception', function () {
       if (requestError) {
         throw requestError;
       }
+    });
+  });
+
+  describe('Request.resourceType', () => {
+    it('should work for document type', async () => {
+      const {page, server} = await getTestState();
+
+      await page.setRequestInterception(true);
+      page.on('request', request => {
+        void request.continue({}, 0);
+      });
+      const response = await page.goto(server.EMPTY_PAGE);
+      const request = response!.request();
+      expect(request.resourceType()).toBe('document');
+    });
+
+    it('should work for stylesheets', async () => {
+      const {page, server} = await getTestState();
+
+      await page.setRequestInterception(true);
+      const cssRequests: HTTPRequest[] = [];
+      page.on('request', request => {
+        if (request.url().endsWith('css')) {
+          cssRequests.push(request);
+        }
+        void request.continue({}, 0);
+      });
+      await page.goto(server.PREFIX + '/one-style.html');
+      expect(cssRequests).toHaveLength(1);
+      const request = cssRequests[0]!;
+      expect(request.url()).toContain('one-style.css');
+      expect(request.resourceType()).toBe('stylesheet');
     });
   });
 });